requirements.txt
aiokafka==0.7.0
amqp==5.0.5
anyjson==0.3.3
asgiref==3.3.1
asyncio==3.4.3
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
billiard==3.6.3.0
cached-property==1.5.2
celery==5.0.5
celery-once==3.0.1
celery-stubs==0.1.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
configparser==5.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
DBUtils==2.0
decorator==4.4.2
distro==1.5.0
Django==3.1.3
django-celery-results==2.0.1
django-crontab==0.7.1
django-rest-framework==0.1.0
django-stubs==1.7.0
djangorestframework==3.12.2
djangorestframework-stubs==1.3.0
docker==4.4.1
docker-compose==1.28.0
dockerpty==0.4.1
docopt==0.6.2
humanize==3.2.0
idna==2.10
importlib-metadata==3.4.0
install==1.3.4
ipython==7.13.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
kafka-python==2.0.2
kazoo==2.5.0
kombu==5.0.2
loguru==0.5.3
MarkupSafe==1.1.1
mypy==0.790
mypy-extensions==0.4.3
paramiko==2.7.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.10
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.4
pykafka==2.8.0
PyMySQL==0.10.1
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.4.1
redis==3.5.3
redis-py-cluster==2.1.0
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tabulate==0.8.7
texttable==1.6.3
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
xlwt==1.3.0
zipp==3.4.0
